get rid of automaton word
