+++ package.json
@@ -9,9 +9,9 @@
     "coverage": "hardhat coverage --testfiles './test/**/*.unit.test.ts'",
     "fork:l1": "hardhat node:fork l1 8545",
     "fork:l2": "hardhat node:fork l2 9545",
-    "deploy-automaton": "ts-node --files ./scripts/optimism/deploy-automaton.ts",
-    "deploy-scratch": "ts-node --files ./scripts/optimism/deploy-scratch.ts",
-    "deploy-steth": "ts-node --files ./scripts/optimism/deploy-new-steth-contracts.ts",
+    "deploy-bridge-without-notifier": "ts-node --files ./scripts/optimism/deploy-bridge-without-notifier.ts",
+    "deploy-bridge": "ts-node --files ./scripts/optimism/deploy-bridge.ts",
+    "deploy-steth": "ts-node --files ./scripts/optimism/deploy-steth.ts",
     "finalize-message": "ts-node --files ./scripts/optimism/finalize-message.ts",
     "test:e2e": "hardhat test ./test/e2e/*.e2e.test.ts",
     "test:managing-e2e": "hardhat test ./test/managing-e2e/*.e2e.test.ts",
